requirements.txt
pandas
openai
langchain[all]
ipython
jupyter
neo4j
py2neo
tiktoken
chromadb
faiss-cpu
scikit-learn
openpyxl
hdbscan
pydantic
networkx
nltk
pgvector
python-dotenv==1.0.1
sqlalchemy